#1901 create a cancel modal component
